several improvements

+++ import_localization.py
@@ -4,7 +4,7 @@ import sys
 from tqdm import tqdm
 from saxonche import PySaxonProcessor
 
-# Run in WSL in /mnt/d/mcmsm as:
+# Run in WSL in /mnt/c/mcmsm as:
 #
 #   python3 import_localization.py <dir>
 #
@@ -36,39 +36,7 @@ print(f"(tested config: saxonche version 12 (saxon/c) with saxon 9.9)")
 #saxon_proc.set_cwd(str(pathlib.Path.cwd()))
 saxon_proc.set_catalog("catalog-wsl.xml")
 xslt_proc = saxon_proc.new_xslt30_processor()
-
-# Ask user for languages
-LANGUAGES = """bg-bg
-cs-cz
-da-dk
-de-de
-el-gr
-es-es
-et-ee
-fi-fi
-fr-fr
-hr-hr
-hu-hu
-it-it
-ja-jp
-ko-kr
-lt-lt
-lv-lv
-nl-nl
-nb-no
-pl-pl
-pt-br
-pt-pt
-ro-ro
-ru-ru
-sk-sk
-sl-si
-sv-se
-th-th
-tr-tr
-vi-vn
-zh-cn
-zh-tw"""
+xslt_exec = xslt_proc.compile_stylesheet(stylesheet_file="import_localization.xslt")
 
 # Collect all source files
 print("Counting files...")
@@ -94,13 +62,15 @@ for translated_file in translated_files:
   localization_file = localization_dir.joinpath(translated_file.name)
 
   # Generate localization file without pretranslated content
-  xslt_proc.transform_to_file(stylesheet_file="import_localization.xslt", 
-    source_file=abspath(translated_file),
+  xslt_exec.transform_to_file(source_file=abspath(translated_file),
     output_file=abspath(localization_file))
-  tqdm.write(f"DITA file imported: {localization_file}")
+  tqdm.write(f"DITA file imported: {localization_file} from {translated_file}")
 
   # Delete link in kit/<language> 
-  kit_file.unlink()
+  if kit_file.exists():
+    kit_file.unlink()
+  else:
+    tqdm.write(f"Warning: kit file missing (would be removed anyway): {kit_file}")
 
 pbar.close()
